pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "erp-system-backend"
version = "0.1.0"
description = "Modular FastAPI backend scaffold with PostgreSQL and AI integration."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "alembic",
  "fastapi",
  "httpx",
  "openai",
  "pydantic<2",
  "psycopg2-binary",
  "psycopg[binary]",
  "sqlalchemy",
  "uvicorn[standard]",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py310"
